ci: split aot-smoke types into separate files + use string.equals
