+++ samples/ZeroAlloc.Serialisation.AotSmoke/MpMessage.cs
@@ -1,0 +1,10 @@
+namespace ZeroAlloc.Serialisation.AotSmoke
+{
+    [MemoryPack.MemoryPackable]
+    [ZeroAllocSerializable(SerializationFormat.MemoryPack)]
+    public sealed partial class MpMessage
+    {
+        public string Id { get; set; } = "";
+        public int Value { get; set; }
+    }
+}
