+++ samples/ZeroAlloc.Serialisation.AotSmoke/Program.cs
@@ -20,32 +20,8 @@ var msgpBuf = new ArrayBufferWriter<byte>();
 new MsgpMessageSerializer().Serialize(msgpBuf, msgp);
 var msgpBack = new MsgpMessageSerializer().Deserialize(msgpBuf.WrittenSpan);
 
-var ok = stjBack?.Id == stj.Id && mpBack?.Id == mp.Id && msgpBack?.Id == msgp.Id;
+var ok = string.Equals(stjBack?.Id, stj.Id, StringComparison.Ordinal)
+    && string.Equals(mpBack?.Id, mp.Id, StringComparison.Ordinal)
+    && string.Equals(msgpBack?.Id, msgp.Id, StringComparison.Ordinal);
 Console.WriteLine(ok ? "AOT smoke: PASS" : "AOT smoke: FAIL");
 return ok ? 0 : 1;
-
-namespace ZeroAlloc.Serialisation.AotSmoke
-{
-    [ZeroAllocSerializable(SerializationFormat.SystemTextJson)]
-    public sealed class StjMessage
-    {
-        public string Id { get; set; } = "";
-        public int Value { get; set; }
-    }
-
-    [MemoryPack.MemoryPackable]
-    [ZeroAllocSerializable(SerializationFormat.MemoryPack)]
-    public sealed partial class MpMessage
-    {
-        public string Id { get; set; } = "";
-        public int Value { get; set; }
-    }
-
-    [MessagePack.MessagePackObject]
-    [ZeroAllocSerializable(SerializationFormat.MessagePack)]
-    public sealed class MsgpMessage
-    {
-        [MessagePack.Key(0)] public string Id { get; set; } = "";
-        [MessagePack.Key(1)] public int Value { get; set; }
-    }
-}
