+++ samples/ZeroAlloc.Serialisation.AotSmoke/StjMessage.cs
@@ -1,0 +1,9 @@
+namespace ZeroAlloc.Serialisation.AotSmoke
+{
+    [ZeroAllocSerializable(SerializationFormat.SystemTextJson)]
+    public sealed class StjMessage
+    {
+        public string Id { get; set; } = "";
+        public int Value { get; set; }
+    }
+}
